+++ CHANGELOG.md
@@ -57,7 +57,8 @@ dj.FreeTable(dj.conn(), "common_session.session_group").drop()
     - Restore #973, allow DLC without position tracking #1100
     - Minor fix to `DLCCentroid` make function order #1112, #1148
     - Pass output path as string to `cv2.VideoWriter` #1150
-    - Set `DLCPosVideo` default processor to `matplotlib` #11XX
+    - Set `DLCPosVideo` default processor to `matplotlib` #1168
+    - Save generated video to temp directory, then move #1168
 
 - Spike Sorting